Rename for ble.h header to avoid name clash.
