+++ nordic-sdk/components/softdevice/common/softdevice_handler/softdevice_handler.c
@@ -43,7 +43,7 @@
 #elif defined(ANT_STACK_SUPPORT_REQD) 
     #include "ant_interface.h"
 #elif defined(BLE_STACK_SUPPORT_REQD)
-    #include "ble.h"
+    #include "nrf_ble.h"
 #endif
